Update the way Dockerfile is versioned.
This feature introduces a new way to version dockerfiles, by
incrementing the number of the BIOCONDUCTOR_PATCH version.

This also introduces a new argument, BIOCONDUCTOR_VERSION. This argument
is needed for "forward" dependencies of images which inhert from the
bioconductor_docker images. If an image inherits from this image, then
it automatically gains access to the bioconductor built binary packages.

The recognition of the image and binary version happens through the
BIOCONDUCTOR_VERSION.